+++ packages/create-svelte/templates/default/src/app.d.ts
@@ -1,5 +1,3 @@
-/// <reference types="@sveltejs/kit" />
-
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 // and what to do when importing types
@@ -9,6 +7,10 @@ declare namespace App {
 	}
 
 	// interface Platform {}
+
+	// interface PrivateEnv {}
+
+	// interface PublicEnv {}
 
 	// interface Session {}