+++ packages/kit/src/runtime/env-public.js
@@ -1,0 +1,7 @@
+/** @type {App.PublicEnv} */
+export let env = {};
+
+/** @type {(environment: Record<string, string>) => void} */
+export function set_public_env(environment) {
+	env = environment;
+}