+++ packages/kit/test/apps/basics/src/routes/errors/stack-trace/index.svelte
@@ -1,0 +1,5 @@
+<script>
+	import bad from './_bad.js';
+</script>
+
+<h1>{bad}</h1>