+++ packages/kit/test/apps/writes/svelte.config.js
@@ -1,0 +1,13 @@
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		prerender: {
+			onError: 'continue'
+		},
+		methodOverride: {
+			allowed: ['PUT', 'PATCH', 'DELETE']
+		}
+	}
+};
+
+export default config;