+++ packages/kit/test/prerendering/trailing-slash/src/routes/standalone-endpoint.json.js
@@ -1,0 +1,7 @@
+export function GET() {
+	return {
+		body: {
+			answer: 42
+		}
+	};
+}